pytorch_toolkit/instance_segmentation/requirements.txt
ansicolors==1.1.8
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
Cython==0.28.2
decorator==4.3.2
future==0.17.1
idna==2.6
kiwisolver==1.0.1
lxml==4.2.3
matplotlib==2.2.2
networkx==2.2
numpy==1.14.3
onnx==1.3.0
opencv-python==3.4.5.20
packaging==17.1
protobuf==3.5.2.post1
py3nvml==0.2.0
pycocotools==2.0.0
pycparser==2.18
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-gflags==3.1.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scipy==1.1.0
six==1.11.0
style==1.1.0
tensorboardX==1.2
termcolor==1.1.0
test-generator==0.1.1
torch==0.4.1
torchvision==0.2.1
tqdm==4.24.0
typing==3.6.4
typing-extensions==3.6.5
update==0.0.1
urllib3==1.22
xmltodict==0.11.0
